+++ lib/need_resume.dart
@@ -1,12 +1,12 @@
 /*
- ----------------------------------------------------------------------
-  NeedResume - Simple package for onResume/viewDidAppear functionality
- ----------------------------------------------------------------------
-  Version: 1.0.2
- ----------------------------------------------------------------------
+ --------------------------------------------------------------------------
+  NeedResume - Simple package for onResume, onPause, onReady functionality
+ --------------------------------------------------------------------------
+  Version: 1.0.3
+ --------------------------------------------------------------------------
   By: Muhammad Faruq Nuruddinsyah
   Copyright 2019. All Rights Reserved.
- ----------------------------------------------------------------------
+ --------------------------------------------------------------------------
 */
 
 import 'package:flutter/material.dart';
@@ -16,15 +16,33 @@ class Resume {
     String source;
 }
 
-abstract class NeedResume {
+abstract class ResumableState<T extends StatefulWidget> extends State with WidgetsBindingObserver {
     Resume resume = new Resume();
+    bool _isPaused = false;
 
     /// Implement your code here
-    void onResume();
+    void onResume() {
+        // TODO: Implement your code here
+    }
+
+    /// Implement your code here
+    void onReady() {
+        // TODO: Implement your code here
+    }
+
+    /// Implement your code here
+    void onPause() {
+        // TODO: Implement your code here
+    }
 
     /// This method is replacement of Navigator.push(), but fires onResume() after route popped
     Future<T> push<T extends Object>(BuildContext context, Route<T> route, [String source]) {
+        _isPaused = true;
+        onPause();
+
         return Navigator.of(context).push(route).then((value) {
+            _isPaused = false;
+
             resume.data = value;
             resume.source = source;
 
@@ -35,12 +53,53 @@ abstract class NeedResume {
 
     /// This method is replacement of Navigator.pushNamed(), but fires onResume() after route popped
     Future<T> pushNamed<T extends Object>(BuildContext context, String routeName, {Object arguments}) {
+        _isPaused = true;
+        onPause();
+
         return Navigator.of(context).pushNamed<T>(routeName, arguments: arguments).then((value) {
+            _isPaused = false;
+
             resume.data = value;
             resume.source = routeName;
 
             onResume();
             return value;
         });
+    }
+
+    @override
+    void initState() {
+        super.initState();
+
+        WidgetsBinding.instance.addObserver(this);
+        WidgetsBinding.instance.addPostFrameCallback((_) => onReady());
+    }
+
+    @override
+    void dispose() {
+        WidgetsBinding.instance.removeObserver(this);
+        super.dispose();
+    }
+
+    @override
+    void didChangeAppLifecycleState(AppLifecycleState state) {
+        switch (state) {
+            case AppLifecycleState.inactive:
+                break;
+            case AppLifecycleState.paused:
+                if (!_isPaused) {
+                    onPause();
+                }
+
+                break;
+            case AppLifecycleState.resumed:
+                if (!_isPaused) {
+                    onResume();
+                }
+
+                break;
+            case AppLifecycleState.suspending:
+                break;
+        }
     }
 }